+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix Dropdown background overlay index
 * #43 Fix Dialog cancel button focus state
 * #50 Add ``page-change`` event on Table
+* #52 Fix Select option
 
 ## 0.2.2